Add 'call' function.
This is very improtant function. It allows to call macro/function with
list of arguments. Now it's possible to write wrappers for variadic
functions/macros. Still need to somehow get rid of recursion in
it though(maybe use progn?).

Also Lips_Invoke now just wraps Lips_Eval. It looks much cleaner
now. Now it accepts only symbols as callable. It is not so bad as it
seems because hash table lookups are very fast.