String allocator: step 1
This is a big one. I finally implemented string/array allocator(arrays
will be supported in future).

We have 16 chunks that store continous bytes of memory. Everytime we
want to allocate a string we lookup if any of the chunks have
available storage for string's data. If we found a chunk we write's to
its contents our string. If not we resize one chunks so it can store
the string.

Step 1 will be merging free region array into chunk's data. We would
store FreeRegion structs alongside strings. This structs will form a
linked list.