Basic garbage collector
Eventually implementing garbage collector was kind of easy. I guess
because we were preparing the code for GC from the start. Next we need
to think about a lot of things:

1. Do we call garbage collector automatically, or the user calls it?
If we call GC automatically then how often do it? (obivously we call
GC when allocating new cells)

2. Should we define a way to 'protect' cells from GC? This is done by
most of GCs, but I don't like. I think we should force users to define
values in environment(assign a value a name).

I also discovered a bug: keys for hashtables are just pointers to
char*, which is not good because strings can move when memory chunks
are out of space. Probably need to replace char* with StringData* in
hast table code.