Reference counting and optimization
Reference counting in strings is finally working: strings are
allocated and freed when needed. But I still think
there's a bug when have more than 1 chunk: for some reason not all
bytes are freed.

Also I implemented a very important optimization: reusing strings for
symbols. Because most symbols in strings are basically the same we can
save a LOT of space by just reusing strings that we're already
allocated. In Lips_CreateSymbolN we just look at all our environments
if we have a same symbol. If yes we reuse it. If no we juts create a
new string.

This optimization won't work for strings: in most cases strings are
unique, so we'll just waste time by trying finding the same string.