Add function 'load' and fix a bug
Finally our repl can load files dynamically. There's also was a
terrifying bug with C functions that call 'progn'. It took me a while
to figure out what's happening. Also I decided that we should throw
exceptions when files are not found.

Now we should decide what file extension we will be using for our
beatiful language. For now I'm sticking with '.lip'. So a lot of files
are called 'lips' haha