GC for plists
Now plists are very nicely builtin in the interpreter. They're very
fast and small, this feature will definetely will be useful. However
we have few things that we need to work on:

1. every plist -> memory allocation. I'm planning to reuse string
allocator for small plists and fallback to malloc when plist is big.

2. there's no builtin iteration over plist. This is a needed
feature. I think we will implement 'for' with recursion for first
time, as non-recursive version will require rebuilding eval loop.

3. There's weird bug described in 'test3_plist.cpp'. I think this
happens because of some hash collision.